style: 修复全 workspace clippy -D warnings 门禁（1640 告警清零）
CI 门禁 cargo clippy --workspace --all-features --all-targets -- -D warnings 从 1640 个告警降至 0（exit 0）：
- 文档类：missing # Errors/# Panics 段落补齐、doc_markdown 反引号（~400 处）
- 宏修复：ExcelRow 派生宏去掉 needless_question_mark（Ok(x?) 包装），消除全 workspace ~350 个告警
- 机械类：long literal 分隔符、field_reassign_with_default、assigning_clones、match_same_arms、items_after_statements 等
- 语义敏感类（float_cmp、cast_*、too_many_arguments/lines、pass_by_value 等 ~60 处）：
  采用 #[allow(clippy::xxx)] + 中文注释，理由为与 Java 语义 1:1 对应（强转/严格比较/签名对齐），不改行为
- 测试断言零改动零削弱；cargo test --workspace --all-features 2771 全绿

+++ easyexcel-core/src/converter/string/string_number_converter.rs
@@ -43,7 +43,7 @@ impl crate::Converter<String> for StringNumberConverter {
             }
             return crate::util::number_utils::format_decimal(
                 &decimal,
-                decimal < BigDecimal::from(0),
+                decimal < 0,
                 Some(pattern),
                 context.column().number_rounding_mode.unwrap_or_default(),
             );
